feat(data-point): refactor entity reducers
reducer-entity -> the new entity instance, reducer-entity-id -> reference to registered entity by id
