+++ packages/data-point/lib/entity-types/base-entity/index.js
@@ -1,6 +1,6 @@
 /*
 NOTE: exporting the resolve function here creates a circular dependency when it's
-imported by reducer-entity/resolve.js -- to avoid this problem, that file imports
+imported by reducer-entity-id/resolve.js -- to avoid this problem, that file imports
 the resolve function directly from base-entity/resolve.js
 */
